feature(pa-72): make errorhandler global, add to actions
