+++ frontend/src/actions/prefAndProfileActions.tsx
@@ -6,6 +6,7 @@ import {Action} from '../types/action'
 import {actionCreator} from '../redux-utils/actionCreator'
 import {operateUserDataRequest} from '../api/HttpClientInstance'
 import {successMessage, errorMessage } from '../constants/errorAndSuccessMessages'
+import { errorHandler } from '../services/ErrorHandler'
 
 // Define action types
 
@@ -60,16 +61,7 @@ export const fetchUserPreferences = () => async (dispatch:Dispatch): Promise<voi
     const result = await operateUserDataRequest.fetchUserPreferences()
     dispatch(setUserPreferences(result))
   } catch (error) {
-    const destructuredMessage = JSON.parse(error.message)
-    if (destructuredMessage) {
-      const [messageArrayFromDestructuredError] = destructuredMessage.errors
-      dispatch(userPreferencesRequestFailed(messageArrayFromDestructuredError.message))
-      const errorText = (messageArrayFromDestructuredError.message).toString()
-      toast.error(errorText)
-    } else {
-      dispatch(userPreferencesRequestFailed(errorMessage.errorUnknown))
-      toast.error(errorMessage.errorUnknown)
-    }
+    errorHandler(error, dispatch)
   }
 }
 
@@ -92,17 +84,7 @@ export const updateUserNickname =
     dispatch(changeUserNickname(result.nickname))
     toast.success(successMessage.successNicknameChange)
   } catch (error) {
-    const destructuredError = {error}
-    const destructuredMessage = JSON.parse(destructuredError.error.message)
-    if (destructuredMessage) {
-    const [messageArrayFromDestructuredError] = destructuredMessage.errors
-    dispatch(userPreferencesRequestFailed(messageArrayFromDestructuredError.message))
-    const errorText = (messageArrayFromDestructuredError.message).toString()
-    toast.error(errorText)
-  } else {
-    dispatch(userPreferencesRequestFailed(errorMessage.errorUnknown))
-    toast.error(errorMessage.errorUnknown)
-  }
+    errorHandler(error, dispatch)
 }
 }
 
@@ -168,15 +150,6 @@ export const updateUserEmail =
     dispatch(changeUserEmailSuccessfull(result.email))
     toast.success(successMessage.successEmailChange)
   } catch (error) {
-    const destructuredError = {error}
-    const destructuredMessage = JSON.parse(destructuredError.error.message)
-    if (destructuredMessage) {
-      const messageArrayFromDestructuredError = destructuredMessage.error
-      dispatch(userPreferencesRequestFailed(messageArrayFromDestructuredError))
-      toast.error(messageArrayFromDestructuredError)
-    } else {
-      dispatch(userPreferencesRequestFailed(errorMessage.errorUnknown))
-      toast.error(errorMessage.errorUnknown)
-    }
+    errorHandler(error, dispatch)
   }
 }
